requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
pillow
openai
streamlit
streamlit-webrtc
ultralytics
numpy
pandas
matplotlib
opencv-python
scipy
requests
opencv-python-headless
plotly


